+++ src/pages/GroupConfig/EditGroup/index.tsx
@@ -8,7 +8,7 @@ import Switcher from "../../../components/Switcher";
 import FormData from "form-data";
 import SimpleToast from "react-native-simple-toast";
 import api from "../../../services/api";
-import Feather from "@expo/vector-icons/Feather";
+import { Feather } from "@expo/vector-icons";
 import { useNavigation, useRoute } from "@react-navigation/native";
 import { Alert } from "react-native";
 import { GroupData } from "../../../../@types/interfaces";